requirements.txt
# Football Predictions Dashboard Requirements
# Production-ready Streamlit application with geospatial visualization

# Core Streamlit and web framework
streamlit>=1.28.0
streamlit-folium>=0.15.0

# Database connectivity
psycopg2-binary>=2.9.7
sqlalchemy>=2.0.0
pandas>=2.1.0

# Visualization libraries
plotly>=5.17.0
pydeck>=0.8.0
folium>=0.15.0

# Data processing
numpy>=1.24.0
datetime

# Additional utilities
python-dotenv>=1.0.0

# Optional: Enhanced mapping capabilities
# mapbox>=3.0.0  # Uncomment if using Mapbox tokens

# Development and debugging (optional)
# streamlit-option-menu>=0.3.6
# streamlit-aggrid>=0.3.4